+++ GUI/pythagore.py
